+++ workers/tests/integrationtests/test_api_indicator_geojson_io.py
@@ -5,7 +5,6 @@ https://fastapi.tiangolo.com/tutorial/testing/
 import os
 import unittest
 from typing import Tuple
-from urllib.parse import urlencode
 
 from fastapi.testclient import TestClient
 from schema import Schema
@@ -23,7 +22,7 @@ from .utils import get_geojson_fixture, oqt_vcr
 class TestApiIndicatorIo(unittest.TestCase):
     def setUp(self):
         self.client = TestClient(app)
-
+        self.endpoint = "/indicator"
         self.indicator_name = "GhsPopComparisonBuildings"
         self.layer_name = "building_count"
 
@@ -41,17 +40,6 @@ class TestApiIndicatorIo(unittest.TestCase):
         schema.validate(geojson_)  # Print information if validation fails
         self.assertTrue(schema.is_valid(geojson_))
 
-    def get_response(self, bpoly):
-        """Return HTTP GET response"""
-        parameters = urlencode(
-            {
-                "name": self.indicator_name,
-                "layerName": self.layer_name,
-                "bpolys": bpoly,
-            }
-        )
-        return self.client.get("/indicator?" + parameters)
-
     def post_response(self, bpoly):
         """Return HTTP POST response"""
         parameters = {
@@ -59,46 +47,38 @@ class TestApiIndicatorIo(unittest.TestCase):
             "bpolys": bpoly,
             "layerName": self.layer_name,
         }
-        return self.client.post("/indicator", json=parameters)
+        return self.client.post(self.endpoint, json=parameters)
 
     @oqt_vcr.use_cassette()
     def test_indicator_bpolys_geometry(self):
         geometry = get_geojson_fixture("heidelberg-altstadt-geometry.geojson")
-        for response in (self.get_response(geometry), self.post_response(geometry)):
-            self.run_tests(response, (self.general_schema, self.feature_schema))
+        response = self.post_response(geometry)
+        self.run_tests(response, (self.general_schema, self.feature_schema))
 
     @oqt_vcr.use_cassette()
     def test_indicator_bpolys_feature(self):
         feature = get_geojson_fixture("heidelberg-altstadt-feature.geojson")
-        for response in (self.get_response(feature), self.post_response(feature)):
-            self.run_tests(response, (self.general_schema, self.feature_schema))
+        response = self.post_response(feature)
+        self.run_tests(response, (self.general_schema, self.feature_schema))
 
     @oqt_vcr.use_cassette()
     def test_indicator_bpolys_featurecollection(self):
         featurecollection = get_geojson_fixture(
             "heidelberg-bahnstadt-bergheim-featurecollection.geojson",
         )
-        for response in (
-            self.get_response(featurecollection),
-            self.post_response(featurecollection),
-        ):
-            self.run_tests(
-                response, (self.general_schema, self.featurecollection_schema)
-            )
-            for feature in response.json()["features"]:
-                self.assertIn("id", feature.keys())
-                self.validate(feature, self.feature_schema)
+        response = self.post_response(featurecollection)
+        self.run_tests(response, (self.general_schema, self.featurecollection_schema))
+        for feature in response.json()["features"]:
+            self.assertIn("id", feature.keys())
+            self.validate(feature, self.feature_schema)
 
     @oqt_vcr.use_cassette()
     def test_indicator_bpolys_size_limit(self):
         feature = get_geojson_fixture("europe.geojson")
-        for response in (
-            self.get_response(feature),
-            self.post_response(feature),
-        ):
-            self.assertEqual(response.status_code, 422)
-            content = response.json()
-            self.assertEqual(content["type"], "SizeRestrictionError")
+        response = self.post_response(feature)
+        self.assertEqual(response.status_code, 422)
+        content = response.json()
+        self.assertEqual(content["type"], "SizeRestrictionError")
 
     @oqt_vcr.use_cassette()
     def test_invalid_set_of_arguments(self):
@@ -115,44 +95,40 @@ class TestApiIndicatorIo(unittest.TestCase):
             "dataset": "foo",
             "featureId": "3",
         }
-
-        for response in (
-            self.client.get("/indicator?" + urlencode(parameters)),
-            self.client.post("/indicator", json=parameters),
-        ):
-            self.assertEqual(response.status_code, 422)
-            content = response.json()
-            self.assertEqual(content["type"], "RequestValidationError")
+        response = self.client.post(self.endpoint, json=parameters)
+        self.assertEqual(response.status_code, 422)
+        content = response.json()
+        self.assertEqual(content["type"], "RequestValidationError")
 
     @oqt_vcr.use_cassette()
     def test_indicator_include_svg(self):
         feature = get_geojson_fixture("heidelberg-altstadt-feature.geojson")
-        url = "/indicator?name={0}&layerName={1}&bpolys={2}&includeSvg={3}".format(
-            self.indicator_name,
-            self.layer_name,
-            feature,
-            True,
-        )
-        response = self.client.get(url)
+        parameters = {
+            "name": self.indicator_name,
+            "layerName": self.layer_name,
+            "bpolys": feature,
+            "includeSvg": True,
+        }
+        response = self.client.post(self.endpoint, json=parameters)
         result = response.json()
         self.assertIn("result.svg", list(result["properties"].keys()))
 
-        url = "/indicator?name={0}&layerName={1}&bpolys={2}&includeSvg={3}".format(
-            self.indicator_name,
-            self.layer_name,
-            feature,
-            False,
-        )
-        response = self.client.get(url)
+        parameters = {
+            "name": self.indicator_name,
+            "layerName": self.layer_name,
+            "bpolys": feature,
+            "includeSvg": False,
+        }
+        response = self.client.post(self.endpoint, json=parameters)
         result = response.json()
         self.assertNotIn("result.svg", list(result["properties"].keys()))
 
-        url = "/indicator?name={0}&layerName={1}&bpolys={2}".format(
-            self.indicator_name,
-            self.layer_name,
-            feature,
-        )
-        response = self.client.get(url)
+        parameters = {
+            "name": self.indicator_name,
+            "layerName": self.layer_name,
+            "bpolys": feature,
+        }
+        response = self.client.post(self.endpoint, json=parameters)
         result = response.json()
         self.assertNotIn("result.svg", list(result["properties"].keys()))