+++ workers/tests/unittests/test_api_request_models.py
@@ -19,7 +19,7 @@ class TestApiRequestModels(unittest.TestCase):
             self.bpolys = json.load(file)
 
     def test_bpolys_valid(self):
-        request_models.BaseRequestModel(bpolys=self.bpolys)
+        request_models.BasePOSTModel(bpolys=self.bpolys)
 
     def test_bpolys_invalid(self):
         bpolys = Polygon(
@@ -27,33 +27,51 @@ class TestApiRequestModels(unittest.TestCase):
         )
 
         with self.assertRaises(ValueError):
-            request_models.BaseRequestModel(bpolys=bpolys)
+            request_models.BasePOSTModel(bpolys=bpolys)
 
     def test_dataset_valid(self):
-        request_models.BaseRequestModel(dataset="regions", feature_id="3")
-        request_models.BaseRequestModel(
+        request_models.BasePOSTModel(dataset="regions", feature_id="3")
+        request_models.BasePOSTModel(
+            dataset="regions", feature_id="3", fid_field="ogc_fid"
+        )
+        request_models.BaseGETModel(dataset="regions", feature_id="3")
+        request_models.BaseGETModel(
             dataset="regions", feature_id="3", fid_field="ogc_fid"
         )
 
     def test_dataset_invalid(self):
         with self.assertRaises(ValueError):
-            request_models.BaseRequestModel(dataset="foo", feature_id="3")
+            request_models.BasePOSTModel(dataset="foo", feature_id="3")
+        with self.assertRaises(ValueError):
+            request_models.BaseGETModel(dataset="foo", feature_id="3")
 
     def test_invalid_set_of_arguments(self):
         with self.assertRaises(ValueError):
-            request_models.BaseRequestModel(
+            request_models.BasePOSTModel(
                 bpolys=self.bpolys, dataset="regions", feature_id="3"
             )
         with self.assertRaises(ValueError):
-            request_models.BaseRequestModel(dataset="regions")
+            request_models.BasePOSTModel(dataset="regions")
         with self.assertRaises(ValueError):
-            request_models.BaseRequestModel(feature_id="3")
+            request_models.BaseGETModel(dataset="regions")
+        with self.assertRaises(ValueError):
+            request_models.BasePOSTModel(feature_id="3")
+        with self.assertRaises(ValueError):
+            request_models.BaseGETModel(feature_id="3")
+        with self.assertRaises(ValueError):
+            request_models.BaseIndicatorModel(name="GhsPopComparisonBuildings")
+        with self.assertRaises(ValueError):
+            request_models.BaseIndicatorModel(layer_name="building_count")
+
+    def test_valid_indicator_layer_combination(self):
+        request_models.BaseIndicatorModel(
+            name="GhsPopComparisonBuildings",
+            layerName="building_count",
+        )
 
     def test_invalid_indicator_layer_combination(self):
         with self.assertRaises(ValueError):
-            request_models.IndicatorRequestModel(
+            request_models.BaseIndicatorModel(
                 name="GhsPopComparisonBuildings",
                 layerName="amenities",
-                dataset="regions",
-                featureId="3",
             )